Trying to split out XML parsing from MindmapNode
