+++ app/build.gradle
@@ -28,4 +28,6 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
     implementation 'androidx.lifecycle:lifecycle-viewmodel:2.0.0'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    compileOnly "org.projectlombok:lombok:1.18.30"
+    annotationProcessor 'org.projectlombok:lombok:1.18.30'
 }
