+++ build.gradle
@@ -7,6 +7,7 @@ buildscript {
     dependencies {
         classpath 'com.android.tools.build:gradle:7.3.1'
         classpath 'com.google.gms:google-services:4.3.10'
+        classpath "org.projectlombok:lombok:1.18.30"
     }
 }
 
@@ -16,3 +17,4 @@ allprojects {
         jcenter()
     }
 }
+
